# Add ClickHouse external data support