--- a/autocomplete/autocomplete.go
+++ b/autocomplete/autocomplete.go
@@ -130,8 +130,16 @@ func (h *Handler) ServeTags(w http.ResponseWriter, r *http.Request) {
 		queryLimit,
 	)
 
-	body, err := clickhouse.Query(scope.WithTable(r.Context(), h.config.ClickHouse.TaggedTable), h.config.ClickHouse.Url, sql,
-		clickhouse.Options{Timeout: h.config.ClickHouse.TreeTimeout.Value(), ConnectTimeout: h.config.ClickHouse.ConnectTimeout.Value()})
+	body, err := clickhouse.Query(
+		scope.WithTable(r.Context(), h.config.ClickHouse.TaggedTable),
+		h.config.ClickHouse.Url,
+		sql,
+		clickhouse.Options{
+			Timeout:        h.config.ClickHouse.TreeTimeout.Value(),
+			ConnectTimeout: h.config.ClickHouse.ConnectTimeout.Value(),
+		},
+		nil,
+	)
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
@@ -228,8 +236,16 @@ func (h *Handler) ServeValues(w http.ResponseWriter, r *http.Request) {
 		limit,
 	)
 
-	body, err := clickhouse.Query(scope.WithTable(r.Context(), h.config.ClickHouse.TaggedTable), h.config.ClickHouse.Url, sql,
-		clickhouse.Options{Timeout: h.config.ClickHouse.IndexTimeout.Value(), ConnectTimeout: h.config.ClickHouse.ConnectTimeout.Value()})
+	body, err := clickhouse.Query(
+		scope.WithTable(r.Context(), h.config.ClickHouse.TaggedTable),
+		h.config.ClickHouse.Url,
+		sql,
+		clickhouse.Options{
+			Timeout:        h.config.ClickHouse.IndexTimeout.Value(),
+			ConnectTimeout: h.config.ClickHouse.ConnectTimeout.Value(),
+		},
+		nil,
+	)
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return