--- a/helper/clickhouse/clickhouse.go
+++ b/helper/clickhouse/clickhouse.go
@@ -163,7 +163,7 @@ func reader(ctx context.Context, dsn string, query string, postBody io.Reader, g
 	if len(queryForLogger) > 500 {
 		queryForLogger = queryForLogger[:395] + "<...>" + queryForLogger[len(queryForLogger)-100:]
 	}
-	logger := scope.Logger(ctx).With(zap.String("query", formatSQL(queryForLogger)), zap.String("request_id", requestID))
+	logger := scope.Logger(ctx).With(zap.String("query", formatSQL(queryForLogger)))
 
 	defer func() {
 		// fmt.Println(time.Since(start), formatSQL(queryForLogger))