--- a/prometheus/handler.go
+++ b/prometheus/handler.go
@@ -75,7 +75,6 @@ func NewHandler(config *config.Config) *Handler {
 func (h *Handler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	logger := scope.Logger(r.Context()).Named("prometheus")
 	r = r.WithContext(scope.WithLogger(r.Context(), logger))
-	w.Header().Add("X-Gch-Request-ID", scope.RequestID(r.Context()))
 	if strings.HasSuffix(r.URL.Path, "/read") {
 		h.read(w, r)
 		return