--- a/render/handler.go
+++ b/render/handler.go
@@ -34,7 +34,6 @@ func NewHandler(config *config.Config) *Handler {
 func (h *Handler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	logger := scope.Logger(r.Context()).Named("render")
 	r = r.WithContext(scope.WithLogger(r.Context(), logger))
-	w.Header().Add("X-Gch-Request-ID", scope.RequestID(r.Context()))
 
 	var prefix string
 	var err error