This should fix #59

Here are benchmark results for old and new logic grouped by the number of metrics returned by the finder:

```
# request.sh:
# curl "http://localhost:8080/render?target=sum($(series "$1"))&from=-24h&until=now&format=json&maxDataPoints=30"

$ for test in 4 320 1432 2366 5876 10151 35957; do hyperfine -w 2 -r 10 "./request.sh $test"; done
Benchmark #1: ./request.sh 4
# old
  Time (mean ± σ):     233.9 ms ±  24.9 ms    [User: 28.6 ms, System: 18.7 ms]
  Range (min … max):   194.1 ms … 264.3 ms    10 runs
# new
  Time (mean ± σ):     215.8 ms ±  19.9 ms    [User: 23.8 ms, System: 15.2 ms]
  Range (min … max):   176.8 ms … 245.4 ms    10 runs

----

Benchmark #1: ./request.sh 320
# old
  Time (mean ± σ):      1.715 s ±  0.756 s    [User: 7.7 ms, System: 4.7 ms]
  Range (min … max):    0.756 s …  3.185 s    10 runs
# new
  Time (mean ± σ):     728.5 ms ±  71.9 ms    [User: 7.3 ms, System: 4.9 ms]
  Range (min … max):   655.3 ms … 871.9 ms    10 runs

----

Benchmark #1: ./request.sh 1432
# old
  Time (mean ± σ):      2.585 s ±  0.521 s    [User: 7.9 ms, System: 2.6 ms]
  Range (min … max):    2.259 s …  4.023 s    10 runs
# new
  Time (mean ± σ):      2.070 s ±  0.267 s    [User: 5.3 ms, System: 4.5 ms]
  Range (min … max):    1.796 s …  2.778 s    10 runs

  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet PC without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.

----

Benchmark #1: ./request.sh 2366
# old
  Time (mean ± σ):      3.847 s ±  1.719 s    [User: 6.4 ms, System: 3.2 ms]
  Range (min … max):    2.699 s …  7.413 s    10 runs

  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet PC without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.
# new
  Time (mean ± σ):      2.745 s ±  0.793 s    [User: 5.6 ms, System: 3.8 ms]
  Range (min … max):    2.264 s …  4.952 s    10 runs

----

Benchmark #1: ./request.sh 5876
# old
  Time (mean ± σ):      6.365 s ±  2.375 s    [User: 5.3 ms, System: 5.0 ms]
  Range (min … max):    4.547 s … 11.924 s    10 runs
# new
  Time (mean ± σ):      4.984 s ±  2.903 s    [User: 7.3 ms, System: 2.7 ms]
  Range (min … max):    3.356 s … 12.362 s    10 runs

  Warning: The first benchmarking run for this command was significantly slower than the rest (12.362 s). This could be caused by (filesystem) caches that were not filled until after the first run. You should consider using the '--warmup' option to fill those caches before the actual benchmark. Alternatively, use the '--prepare' option to clear the caches before each timing run.

----

Benchmark #1: ./request.sh 10151
# old
  Time (mean ± σ):     19.780 s ±  4.474 s    [User: 8.7 ms, System: 5.8 ms]
  Range (min … max):   16.728 s … 31.647 s    10 runs
# new
  Time (mean ± σ):     11.199 s ±  3.898 s    [User: 5.7 ms, System: 5.7 ms]
  Range (min … max):    8.273 s … 19.967 s    10 runs

----

Benchmark #1: ./request.sh 35957 # is failed because of too long queries, more than 512k characters
# old
  Time (mean ± σ):     725.6 ms ± 142.7 ms    [User: 14.4 ms, System: 17.1 ms]
  Range (min … max):   548.4 ms … 1021.2 ms    10 runs
# new
  Time (mean ± σ):     13.846 s ±  5.729 s    [User: 6.7 ms, System: 5.0 ms]
  Range (min … max):   11.167 s … 30.026 s    10 runs

  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet PC without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.
```