@Felixoid 2020-10-15
If somebody interested in profiles I've done during the mentioned benchmarks, here they are
[graphite-clickhouse.samples.cpu.pb.gz](https://github.com/lomik/graphite-clickhouse/files/5384913/graphite-clickhouse.samples.cpu.pb.gz)
[graphite-clickhouse_in.samples.cpu.pb.gz](https://github.com/lomik/graphite-clickhouse/files/5384914/graphite-clickhouse_in.samples.cpu.pb.gz)