@Felixoid 2020-10-26
This works in our production already for a week, should be safe to use